requirements.txt
aiohttp==3.11.12
alibabacloud_fc_open20210406==2.0.11
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
Flask==3.1.0
matplotlib==3.10.0
numpy==2.2.3
oss2==2.19.1
packaging==24.2
pydantic==1.10.8
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
redis==5.2.1
Requests==2.32.3
setuptools==66.1.1
